Add support for upload directly to static's root
This commit adds an option 'AWS_STATIC_DESTINATION_TYPE' that chooses a
strategy of where to upload files to S3. It's currently two options,
'STATIC_ROOT' and the unnamed default.

If STATIC_ROOT is set, files are uploaded to the folder specified by the
AWS_STATIC_PREFIX option. The rationale behind this change is to let
browser clients keep the cache of unchanged files. on each deploy.
However, using the STATIC_ROOT strategy, the files should be hashed
before upload - otherwise the upload will overwrite existing files.